refactor(session): split role-specific services
